+++ hive/lib/src/backend/js/native/backend_manager.dart
@@ -18,6 +18,7 @@ class BackendManager implements BackendManagerInterface {
     String? path,
     bool crashRecovery,
     HiveCipher? cipher,
+    int? keyCrc,
     String? collection,
   ) async {
     // compatibility for old store format