+++ hive/lib/src/backend/storage_backend.dart
@@ -54,6 +54,7 @@ abstract class BackendManagerInterface {
     String? path,
     bool crashRecovery,
     HiveCipher? cipher,
+    int? keyCrc,
     String? collection,
   );