+++ hive/lib/src/backend/storage_backend_memory.dart
@@ -9,13 +9,14 @@ import 'package:hive_ce/src/box/keystore.dart';
 /// In-memory Storage backend
 class StorageBackendMemory extends StorageBackend {
   final HiveCipher? _cipher;
+  final int? _keyCrc;
 
   final FrameHelper _frameHelper;
 
   Uint8List? _bytes;
 
   /// Not part of public API
-  StorageBackendMemory(Uint8List? bytes, this._cipher)
+  StorageBackendMemory(Uint8List? bytes, this._cipher, this._keyCrc)
       : _bytes = bytes,
         _frameHelper = FrameHelper();
 
@@ -37,6 +38,7 @@ class StorageBackendMemory extends StorageBackend {
       keystore,
       registry,
       _cipher,
+      _keyCrc,
     );
 
     if (recoveryOffset != -1) {