+++ hive/lib/src/isolate/isolated_hive_impl/hive_isolate.dart
@@ -7,21 +7,6 @@ import 'package:meta/meta.dart';
 ///
 /// Used for testing
 abstract class HiveIsolate {
-  /// Warning message printed when using [IsolatedHive] without an [IsolateNameServer]
-  static final noIsolateNameServerWarning = '''
-⚠️ WARNING: HIVE MULTI-ISOLATE RISK DETECTED ⚠️
-
-Using IsolatedHive without an IsolateNameServer is unsafe. This can lead to
-DATA CORRUPTION as Hive boxes are not designed for concurrent access across
-isolates. Using an IsolateNameServer allows IsolatedHive to maintain a single
-isolate for all Hive operations.
-
-RECOMMENDED ACTIONS:
-- Initialize IsolatedHive with IsolatedHive.initFlutter from hive_ce_flutter
-- Provide your own IsolateNameServer
-
-''';
-
   /// Access to the isolate connection for testing
   @visibleForTesting
   IsolateConnection get connection;